Add descriptions for configuration options
