+++ docs/topics/formats/html.md
@@ -22,7 +22,8 @@ This format comes standard in all runners:
 
 ## Configuration
 
-HTML format is Dokka's base format, so it is configurable through `DokkaBase` and `DokkaBaseConfiguration`:
+HTML format is Dokka's base format, so it is configurable through `DokkaBase` and `DokkaBaseConfiguration`
+classes:
 
 <tabs group="build-script">
 <tab title="Kotlin" group-key="kotlin">
@@ -160,7 +161,7 @@ Via [JSON configuration](cli.md#running-with-json-configuration):
 </tab>
 </tabs>
 
-For details, see [configuring Dokka plugins](plugins_introduction.md#configuring-dokka-plugins) section.
+For more details, see [configuring Dokka plugins](plugins_introduction.md#configuring-dokka-plugins) topic.
 
 ### Configuration options
 
@@ -176,7 +177,7 @@ For details, see [configuring Dokka plugins](plugins_introduction.md#configuring
 
 ## Customization
 
-HTML format is highly customizable to the point where you wouldn't be able to tell it was generated by Dokka.
+HTML format provides a number of customization options.
 
 ### Customizing styles
 
@@ -192,7 +193,7 @@ Files with the same name will be overwritten, so it is possible to override styl
 | `prism.css`          | Styles for [PrismJS](https://prismjs.com/) syntax highlighter |
 | `jetbrains-mono.css` | Font styling                                                  |
 
-All of these are 
+Source code for all of the used styles is 
 [available on GitHub](https://github.com/Kotlin/dokka/tree/%dokkaVersion%/plugins/base/src/main/resources/dokka/styles).
 
 ### Customizing assets
@@ -203,7 +204,7 @@ You can provide your own images to be bundled with documentation by setting `cus
 Files with the same name will be overwritten, so it is possible to override images and icons that Dokka uses. The most
 useful and relevant one being `logo-icon.svg`, which is the image that's used in the header. The rest is mostly icons.
 
-You can view all images used by Dokka on 
+You can find all images used by Dokka on 
 [GitHub](https://github.com/Kotlin/dokka/tree/%dokkaVersion%/plugins/base/src/main/resources/dokka/images).
 
 ### Changing the logo
@@ -211,10 +212,10 @@ You can view all images used by Dokka on
 To customize the logo, you can begin by [providing your own asset](#customizing-assets) for `logo-icon.svg`. If your
 image has similar size, it should not look out of place.
 
-However, if your image has different dimensions or you want to use a `.png` image instead of the default `.svg`,
-you can [override `logo-styles.css` stylesheet](#customizing-styles).  
+However, if your image has different dimensions or you want to use a `.png` image instead of the default `.svg` file,
+you can [override `logo-styles.css` stylesheet](#customizing-styles) and make it fit.  
 
-For more details, see 
+For an example of how to do it, see 
 [custom format example project](https://github.com/Kotlin/dokka/tree/%dokkaVersion%/examples/gradle/dokka-customFormat-example).
 
 ### Modifying footer
@@ -226,7 +227,7 @@ You can modify the footer message by setting `footerMessage` [configuration prop
 Dokka provides the ability to modify [FreeMarker](https://freemarker.apache.org/) templates used for generating 
 documentation pages.
 
-You can completely change the header, add your own banners/menus/search, load analytics, change body styling and so on.
+You can change the header completely, add your own banners/menus/search, load analytics, change body styling and so on.
 
 Dokka uses the following templates:
 
@@ -238,7 +239,7 @@ Dokka uses the following templates:
 | `includes/page_metadata.ft`       | Metadata used within `<head>` container.                                                                          |
 | `includes/source_set_selector.ft` | [Source set](https://kotlinlang.org/docs/multiplatform-discover-project.html#source-sets) selector in the header. |
 
-`base.ftl` template is the base template and it includes all the other ones. You can find all templates 
+`base.ftl` template is the base template and it includes all the other ones. You can find source code for all templates 
 [on GitHub](https://github.com/Kotlin/dokka/tree/%dokkaVersion%/plugins/base/src/main/resources/dokka/templates).
 
 You can override any template by setting `templatesDir` [configuration property](#configuration). Dokka will look
@@ -268,11 +269,11 @@ context and thus need to be resolved at later stages by the [MultiModule](gradle
 
 #### Directives
 
-Dokka-defined [directives](https://freemarker.apache.org/docs/ref_directive_userDefined.html) can also be used:
+You can also use the following Dokka-defined [directives](https://freemarker.apache.org/docs/ref_directive_userDefined.html):
 
-| **Variable** | **Description**                                                                                                                        |
-|----------------------------|----------------------------------------------------------------------------------------------------------------------------------------|
-| `<@content/>`              | Main page content.                                                                                                                     |
-| `<@resources/>`            | Resources such as scripts and stylesheets.                                                                                             |
-| `<@version/>`              | Module version taken from configuration. If [versioning plugin](versioning.md) is applied, it will be replaced with version navigator. |
+| **Variable**    | **Description**                                                                                                                        |
+|-----------------|----------------------------------------------------------------------------------------------------------------------------------------|
+| `<@content/>`   | Main page content.                                                                                                                     |
+| `<@resources/>` | Resources such as scripts and stylesheets.                                                                                             |
+| `<@version/>`   | Module version taken from configuration. If [versioning plugin](versioning.md) is applied, it will be replaced with version navigator. |
