+++ runners/gradle-plugin/src/main/kotlin/org/jetbrains/dokka/gradle/AbstractDokkaTask.kt
@@ -23,48 +23,160 @@ import kotlin.reflect.full.createInstance
 @DisableCachingByDefault(because = "Abstract super-class, not to be instantiated directly")
 abstract class AbstractDokkaTask : DefaultTask() {
 
+    /**
+     * Display name used to refer to the module. Used for ToC, navigation, logging, etc.
+     *
+     * If set for a single-project build or a MultiModule task, will be used as project name.
+     *
+     * Default is Gradle project name.
+     */
     @Input
     val moduleName: Property<String> = project.objects.safeProperty<String>()
         .safeConvention(project.name)
 
+    /**
+     * Module version.
+     *
+     * If set for a single-project build or a MultiModule task, will be used
+     * as project version by the versioning plugin.
+     *
+     * Default is Gradle project version.
+     */
     @Input
     val moduleVersion: Property<String> = project.objects.safeProperty<String>()
         .safeConvention(project.version.toString())
 
+    /**
+     * Directory to which documentation will be generated, regardless of format.
+     * Can be set on per-task basis.
+     *
+     * Default is `project/buildDir/taskName.removePrefix("dokka").decapitalize()`, so
+     * for `dokkaHtmlMultiModule` task it will be `project/buildDir/htmlMultiModule`
+     */
     @OutputDirectory
     val outputDirectory: Property<File> = project.objects.safeProperty<File>()
         .safeConvention(project.provider { defaultDokkaOutputDirectory() })
+
+    /**
+     * Configuration for Dokka plugins. This property is not expected to be used directly - if possible, use
+     * [pluginConfiguration] blocks (preferred) or [pluginsMapConfiguration] instead.
+     */
+    @Input
+    val pluginsConfiguration: ListProperty<in DokkaConfiguration.PluginConfiguration> = project.objects.listProperty()
+
+    /**
+     * JSON configuration of Dokka plugins.
+     *
+     * Key is fully qualified Dokka plugin name, value is its configuration in JSON.
+     *
+     * Example:
+     *
+     * ```kotlin
+     * tasks.dokkaHtml {
+     *     val dokkaBaseConfiguration = """
+     *     {
+     *         "customAssets": ["${file("assets/my-image.png")}"],
+     *         "customStyleSheets": ["${file("assets/my-styles.css")}"],
+     *         "footerMessage": "(c) 2022 MyOrg"
+     *     }
+     *     """
+     *     pluginsMapConfiguration.set(
+     *         mapOf("org.jetbrains.dokka.base.DokkaBase" to dokkaBaseConfiguration)
+     *     )
+     * }
+     * ```
+     */
+    @Input
+    val pluginsMapConfiguration: MapProperty<String, String> = project.objects.mapProperty()
+
+    /**
+     * Whether to suppress obvious functions.
+     *
+     * A function is considered to be obvious if it is:
+     * - Inherited from `kotlin.Any`, `Kotlin.Enum`, `java.lang.Object` or `java.lang.Enum`,
+     *   such as `equals`, `hashCode`, `toString`.
+     * - Synthetic (generated by the compiler) and does not have any documentation, such as
+     *   `dataClass.componentN` or `dataClass.copy`.
+     *
+     * Default is `true`
+     */
+    @Input
+    val suppressObviousFunctions: Property<Boolean> = project.objects.safeProperty<Boolean>()
+        .safeConvention(DokkaDefaults.suppressObviousFunctions)
+
+    /**
+     * Whether to suppress inherited members that aren't explicitly overridden in a given class.
+     *
+     * Note: this can suppress functions such as `equals`/`hashCode`/`toString`, but cannot suppress
+     * synthetic functions such as `dataClass.componentN` and `dataClass.copy`. Use [suppressObviousFunctions]
+     * for that.
+     *
+     * Default is `false`.
+     */
+    @Input
+    val suppressInheritedMembers: Property<Boolean> = project.objects.safeProperty<Boolean>()
+        .safeConvention(DokkaDefaults.suppressInheritedMembers)
+
+    /**
+     * Whether to resolve remote files/links over network.
+     *
+     * This includes package-lists used for generating external documentation links:
+     * for instance, to make classes from standard library clickable.
+     *
+     * Setting this to `true` can significantly speed up build times in certain cases,
+     * but can also worsen documentation quality and user experience, for instance by
+     * not resolving some dependency's class/member links.
+     *
+     * When using offline mode, you can cache fetched files locally and provide them to
+     * Dokka as local paths. For instance, see [GradleExternalDocumentationLinkBuilder].
+     *
+     * Default is `false`.
+     */
+    @Input
+    val offlineMode: Property<Boolean> = project.objects.safeProperty<Boolean>()
+        .safeConvention(DokkaDefaults.offlineMode)
+
+    /**
+     * Whether to fail documentation generation if Dokka has emitted a warning or an error.
+     * Will wait until all errors and warnings have been emitted first.
+     *
+     * This setting works well with [GradleDokkaSourceSetBuilder.reportUndocumented]
+     *
+     * Default is `false`.
+     */
+    @Input
+    val failOnWarning: Property<Boolean> = project.objects.safeProperty<Boolean>()
+        .safeConvention(DokkaDefaults.failOnWarning)
 
     @Optional
     @InputDirectory
     @PathSensitive(PathSensitivity.RELATIVE)
     val cacheRoot: Property<File?> = project.objects.safeProperty()
 
-    @Input
-    val failOnWarning: Property<Boolean> = project.objects.safeProperty<Boolean>()
-        .safeConvention(DokkaDefaults.failOnWarning)
-
-    @Input
-    val suppressObviousFunctions: Property<Boolean> = project.objects.safeProperty<Boolean>()
-        .safeConvention(DokkaDefaults.suppressObviousFunctions)
-
-    @Input
-    val suppressInheritedMembers: Property<Boolean> = project.objects.safeProperty<Boolean>()
-        .safeConvention(DokkaDefaults.suppressInheritedMembers)
-
-    @Input
-    val offlineMode: Property<Boolean> = project.objects.safeProperty<Boolean>()
-        .safeConvention(DokkaDefaults.offlineMode)
-
-    @Input
-    val pluginsConfiguration: ListProperty<in DokkaConfiguration.PluginConfiguration> = project.objects.listProperty()
-
     /**
-     * Used to keep compatibility with gradle using Kotlin lower than 1.3.50
+     * Type-safe configuration for a Dokka plugin.
+     *
+     * Note: this is available in Kotlin DSL only, if Dokka Gradle plugin was applied through `plugins` block
+     * and the configured plugin can be found on classpath, which may require adding a classpath dependency
+     * to `buildscript` block in case of external plugins. Some Dokka plugins, such as
+     * [org.jetbrains.dokka.base.DokkaBase], are on classpath by default.
+     *
+     * Example:
+     *
+     * ```kotlin
+     * import org.jetbrains.dokka.base.DokkaBase
+     * import org.jetbrains.dokka.base.DokkaBaseConfiguration
+     *
+     * tasks.dokkaHtml {
+     *     pluginConfiguration<DokkaBase, DokkaBaseConfiguration> {
+     *         footerMessage = "Test"
+     *     }
+     * }
+     * ```
+     *
+     * @param P Plugin class that extends [DokkaPlugin]
+     * @param T Plugin configuration class that extends [ConfigurableBlock]
      */
-    @Input
-    val pluginsMapConfiguration: MapProperty<String, String> = project.objects.mapProperty()
-
     inline fun <reified P : DokkaPlugin, reified T : ConfigurableBlock> pluginConfiguration(block: T.() -> Unit) {
         val instance = T::class.createInstance().apply(block)
         val pluginConfiguration = PluginConfigurationImpl(
